fix native tests for real
